(hel-445) move tagmultiniveau dans entite juridiqueé g
